+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 ### Fixed
 
 - [SQLite Dialect] Fix Sqlite 3.18 missing functions (#5759 by by [Griffio][griffio])
+- [Compiler] Keep track of table aliases when expanding SELECT * statements (#5816 by [Jon Poulton][jonapoul])
 
 
 ## [2.1.0] - 2025-05-16